requirements.txt
pytelegrambotapi==4.0.0
fs-moysklad-api==0.5.1
requests~=2.24.0
boto3~=1.18.38